requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
mcp>=1.0.0
pymssql>=2.2.0
pyyaml>=6.0
boto3>=1.35.0
